back/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
base58==2.1.1
bitarray==2.7.3
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
cytoolz==0.12.1
distlib==0.3.6
Django==4.1.7
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
fastapi==0.95.1
filelock==3.8.2
fonttools==4.39.3
frozenlist==1.3.3
h11==0.14.0
hexbytes==0.3.0
idna==3.4
importlib-resources==5.12.0
ipfshttpclient==0.8.0a2
jsonschema==4.17.3
kiwisolver==1.4.4
lru-dict==1.1.8
matplotlib==3.7.1
multiaddr==0.0.9
multidict==6.0.4
mypy-extensions==1.0.0
Naked==0.1.32
netaddr==0.8.0
numpy==1.24.3
packaging==23.1
parsimonious==0.8.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==2.6.0
protobuf==3.19.5
pycryptodome==3.16.0
pycryptodome-test-vectors==1.0.11
pycryptodomex==3.16.0
pydantic==1.10.7
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pywin32==306
PyYAML==6.0
requests==2.28.2
rlp==2.0.1
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
starlette==0.26.1
tomli==2.0.1
toolz==0.12.0
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
uvicorn==0.22.0
varint==1.0.2
virtualenv==20.17.1
web3==5.31.3
websockets==9.1
yarl==1.9.2
zipp==3.15.0